v.0.2

+++ ft_atoi.c
@@ -6,7 +6,7 @@
 /*   By: mmonahan <mmonahan@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/12/12 17:16:58 by mmonahan          #+#    #+#             */
-/*   Updated: 2019/01/16 20:56:38 by mmonahan         ###   ########.fr       */
+/*   Updated: 2019/01/17 16:59:55 by mmonahan         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int	ft_atoi(const char *str)
 {
-	unsigned long int	tmp;
+	unsigned long long	tmp;
 	int					minus;
 
 	tmp = 0;
